client management mass assign project with select and creation,refactor some code,fix bugs where the array is not present after an ajax call

+++ app/Http/Controllers/User/DeleteClient.php
@@ -4,10 +4,13 @@ namespace App\Http\Controllers\User;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller as BaseController;
-use App\Client;
 
 class DeleteClient extends BaseController
 {
+    
+    protected $message;
+
+    protected $code = 204;
     /**
      * Create a new controller instance.
      *
@@ -25,6 +28,16 @@ class DeleteClient extends BaseController
      */
     public function __invoke($client)
     {
-        $client->delete();
+        if($this->getAuth()->id === $this->getTenant()->id){
+            $projects = $client->projects;
+            foreach($projects as $project){
+                $project->byClient()->dissociate();
+            }
+            $client->delete();
+            return response()->json([], $this->code);
+        }
+        $this->message = 'UnAuthorized Action';
+        $this->code = 401;
+        return response()->json(['message' => $this->message],$this->code);
     }
 }
